int has been replaced with size_t.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2445 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUTools++/TU/Geometry++.h
@@ -1734,7 +1734,7 @@ BoundingBox<P>::clear()
 template <class P> template <class T_, size_t D_> BoundingBox<P>&
 BoundingBox<P>::expand(const Array<T_, D_>& p)
 {
-    for (int i = 0; i < dim(); ++i)
+    for (size_t i = 0; i < dim(); ++i)
     {
 	_min[i] = std::min(_min[i], p[i]);
 	_max[i] = std::max(_max[i], p[i]);
